Implement macro logic
This required pulling out the common types used by the proc macro into a
separate library. I had intended to do this eventually anyway, but it
turned out to be necessary to prevent libclang from interfering with the
LLVM configuration of the compiler due to being pulled in by the proc
macro!